Updating scoping of migrations
